refactor: move run index, run dates, external-pid probe and find_children to data layer
Kills all six data->services imports. Baseline 34 -> 28.
