+++ src/quodeq/data/sqlite/_index_sync.py
