+++ tests/data/test_index_sync.py
@@ -12,8 +12,8 @@ from pathlib import Path
 import pytest
 
 from quodeq.shared.run_status import RunState, write_status
-from quodeq.services.run_index import open_index, sync_index
-from quodeq.services._index_sync import (
+from quodeq.data.sqlite.run_index import open_index, sync_index
+from quodeq.data.sqlite._index_sync import (
     _is_pid_alive,
     _sync_legacy_run,
     _upsert_from_status,
