+++ tests/services/test_cancel_discard_purge.py
@@ -265,7 +265,7 @@ class TestProviderDiscardPurgesRun:
         # the liveness check report the pid dead once "killed" — but leave
         # status.json untouched, exactly like a wedged process would.
         import quodeq.services._external_jobs as _ext_mod
-        import quodeq.services._index_sync as _sync_mod
+        import quodeq.data.sqlite._index_sync as _sync_mod
         original_kill_tree = _ext_mod._kill_tree
         original_alive = _sync_mod._is_pid_alive
         pid_killed = False
